+++ pages/_app.js
@@ -1,4 +1,4 @@
-import '../styles/global.css'
+import '../styles/global.scss'
 
 export default function App({ Component, pageProps }) {
   return <Component {...pageProps} />
