+++ pages/index.js
@@ -2,7 +2,7 @@ import Head from 'next/head'
 import Link from 'next/link'
 import Layout, { siteTitle } from '../components/layout'
 import Hero from '../components/hero'
-import utilStyles from '../styles/utils.module.css'
+import utilStyles from '../styles/utils.module.scss'
 
 export default function Home() {
   return (
